requirements.txt
fastapi
streamlit
uvicorn
jinja2
python-dotenv
requests
langchain
langchain-core
langgraph
groq
langchain-groq
langchain-community
langchain-text-splitters
langchain-openai
langchain-pinecone
langchain-ollama
supabase
langchain-experimental
psycopg2-binary